+++ lib/rdf/quad.ex
@@ -11,7 +11,7 @@ defmodule RDF.Quad do
   @doc """
   Creates a `RDF.Quad` with proper RDF values.
 
-  An error is raised when the given elements are not convertible to RDF values.
+  An error is raised when the given elements are not coercible to RDF values.
 
   Note: The `RDF.quad` function is a shortcut to this function.
 
@@ -24,17 +24,17 @@ defmodule RDF.Quad do
   """
   def new(subject, predicate, object, graph_context) do
     {
-      Statement.convert_subject(subject),
-      Statement.convert_predicate(predicate),
-      Statement.convert_object(object),
-      Statement.convert_graph_name(graph_context)
+      Statement.coerce_subject(subject),
+      Statement.coerce_predicate(predicate),
+      Statement.coerce_object(object),
+      Statement.coerce_graph_name(graph_context)
     }
   end
 
   @doc """
   Creates a `RDF.Quad` with proper RDF values.
 
-  An error is raised when the given elements are not convertible to RDF values.
+  An error is raised when the given elements are not coercible to RDF values.
 
   Note: The `RDF.quad` function is a shortcut to this function.