+++ lib/rdf/triple.ex
@@ -11,7 +11,7 @@ defmodule RDF.Triple do
   @doc """
   Creates a `RDF.Triple` with proper RDF values.
 
-  An error is raised when the given elements are not convertible to RDF values.
+  An error is raised when the given elements are not coercible to RDF values.
 
   Note: The `RDF.triple` function is a shortcut to this function.
 
@@ -24,16 +24,16 @@ defmodule RDF.Triple do
   """
   def new(subject, predicate, object) do
     {
-      Statement.convert_subject(subject),
-      Statement.convert_predicate(predicate),
-      Statement.convert_object(object)
+      Statement.coerce_subject(subject),
+      Statement.coerce_predicate(predicate),
+      Statement.coerce_object(object)
     }
   end
 
   @doc """
   Creates a `RDF.Triple` with proper RDF values.
 
-  An error is raised when the given elements are not convertible to RDF values.
+  An error is raised when the given elements are not coercible to RDF values.
 
   Note: The `RDF.triple` function is a shortcut to this function.