+++ test/unit/description_test.exs
@@ -54,7 +54,7 @@ defmodule RDF.DescriptionTest do
       assert description_includes_predication(desc2, {EX.predicate, uri(EX.Object)})
     end
 
-    test "from a map with convertible RDF term" do
+    test "from a map with coercible RDF term" do
       desc = Description.new(EX.Subject, %{EX.Predicate => EX.Object})
       assert description_of_subject(desc, uri(EX.Subject))
       assert description_includes_predication(desc, {uri(EX.Predicate), uri(EX.Object)})
@@ -79,7 +79,7 @@ defmodule RDF.DescriptionTest do
         |> description_includes_predication({EX.predicate, uri(EX.Object)})
     end
 
-    test "a predicate-object-pair of convertible RDF terms" do
+    test "a predicate-object-pair of coercible RDF terms" do
       assert Description.add(description(),
               "http://example.com/predicate", uri(EX.Object))
         |> description_includes_predication({EX.predicate, uri(EX.Object)})
@@ -161,7 +161,7 @@ defmodule RDF.DescriptionTest do
       assert description_includes_predication(desc, {EX.predicate1, uri(EX.Object4)})
     end
 
-    test "a map of predications with convertible RDF terms" do
+    test "a map of predications with coercible RDF terms" do
       desc = description([{EX.predicate1, EX.Object1}, {EX.predicate2, EX.Object2}])
         |> Description.add(%{EX.predicate3 => EX.Object3})
 
@@ -181,7 +181,7 @@ defmodule RDF.DescriptionTest do
       assert description_includes_predication(desc, {EX.predicate3, bnode(:foo)})
     end
 
-    test "a map of predications with inconvertible RDF terms" do
+    test "a map of predications with non-coercible RDF terms" do
       assert_raise RDF.InvalidURIError, fn ->
         Description.add(description(), %{"not a URI" => uri(EX.Object)})
       end
@@ -200,7 +200,7 @@ defmodule RDF.DescriptionTest do
       assert Description.add(desc, {EX.predicate, literal(42)}) == desc
     end
 
-    test "non-convertible Triple elements are causing an error" do
+    test "non-coercible Triple elements are causing an error" do
       assert_raise RDF.InvalidURIError, fn ->
         Description.add(description(), {"not a URI", uri(EX.Object)})
       end